newSpaceInvaders/requirements.txt
actions==1.0.3
bunch==1.0.1
PyYAML==5.1.2
